# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Aarambh NextToppers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface YouTubeVideo {
  id: string;
  title: string;
  description: string;
  thumbnail: string;
  publishedAt: string;
  channelTitle: string;
  /** @nullable */
  viewCount?: string | null;
  /** @nullable */
  duration?: string | null;
}

export interface ApiError {
  error: string;
}

export type GetChannelVideosParams = {
channelId: string;
maxResults?: number;
};

export type SearchYouTubeVideosParams = {
channelId: string;
query?: string;
maxResults?: number;
};