# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Aarambh NextToppers
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: youtube
    description: YouTube proxy operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /youtube/channel-videos:
    get:
      operationId: getChannelVideos
      tags: [youtube]
      summary: Get videos from a YouTube channel
      description: Proxies YouTube Data API to fetch latest videos from a channel
      parameters:
        - name: channelId
          in: query
          required: true
          schema:
            type: string
        - name: maxResults
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YouTubeVideo"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /youtube/search:
    get:
      operationId: searchYouTubeVideos
      tags: [youtube]
      summary: Search YouTube videos by channel
      description: Search for videos within a YouTube channel
      parameters:
        - name: channelId
          in: query
          required: true
          schema:
            type: string
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: maxResults
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YouTubeVideo"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    YouTubeVideo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        publishedAt:
          type: string
        channelTitle:
          type: string
        viewCount:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
      required:
        - id
        - title
        - description
        - thumbnail
        - publishedAt
        - channelTitle
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error