# lib/api-zod/src/generated/types/youTubeVideo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Aarambh NextToppers
 * OpenAPI spec version: 0.1.0
 */

export interface YouTubeVideo {
  id: string;
  title: string;
  description: string;
  thumbnail: string;
  publishedAt: string;
  channelTitle: string;
  /** @nullable */
  viewCount?: string | null;
  /** @nullable */
  duration?: string | null;
}